refactor(networks): drop FastSampler, merge scale-free into msm.py
Drops _fast_sampler.py and scalefree.py; MSMScaleFreeNetwork now lives
alongside the other MSM classes. Selection sampling uses an inline cumsum
+ searchsorted (same algorithm FastSampler implemented). ss.choice covers
the same need at L1 — no upstream candidate.

Adds a CRN caveat to the class docstring: the kernel uses global weighted-
categorical draws over a pair catalog, which is intrinsically not branching-
stable under starsim CRN. Same precedent as the desired_age_bucket matcher.

Adds a uniform-λ fallback in _build_kernel: when the Whittles formula
collapses (no biological heterogeneity), structure comes from the mix-
weights hook alone with a constant-rate baseline.
